WTF-372 Create the 'modals' package to hold modal code
